First shot at full model
It’s compiled and it’s running. We’ll see if it finishes and if the
results make sense.

The trick, if it works, is simply to subset the regression parameters
(slopes and covariance matrix) to the responses that are relevant after
having divided the data into complete and partial.