Corrected error in path analysis code.
Also trying environmental PCAs as covariates in analysis of garden data.

Path analysis error in multivariate model for leaf. Should have only a
single mean for each leaf trait instead of a mean for each individual
and each leaf trait. We don’t have a regression for the leaf traits so
all individuals are, a priori, expected to have the same trait value.